Rename payment to gateway, and order to payment
