+++ src/Payum/Bridge/JMSPayment/JmsGatewayFactory.php
@@ -4,15 +4,15 @@ namespace Payum\Bridge\JMSPayment;
 use Payum\Bridge\JMSPayment\Action\CaptureAction;
 use Payum\Bridge\JMSPayment\Action\StatusAction;
 use Payum\Core\Bridge\Spl\ArrayObject;
-use Payum\Core\PaymentFactoryInterface;
-use Payum\Core\PaymentFactory as CorePaymentFactory;
+use Payum\Core\GatewayFactoryInterface;
+use Payum\Core\GatewayFactory as CoreGatewayFactory;
 
-class PaymentFactory implements PaymentFactoryInterface
+class JmsGatewayFactory implements GatewayFactoryInterface
 {
     /**
-     * @var PaymentFactoryInterface
+     * @var GatewayFactoryInterface
      */
-    protected $corePaymentFactory;
+    protected $coreGatewayFactory;
 
     /**
      * @var array
@@ -21,11 +21,11 @@ class PaymentFactory implements PaymentFactoryInterface
 
     /**
      * @param array $defaultConfig
-     * @param PaymentFactoryInterface $corePaymentFactory
+     * @param GatewayFactoryInterface $coreGatewayFactory
      */
-    public function __construct(array $defaultConfig = array(), PaymentFactoryInterface $corePaymentFactory = null)
+    public function __construct(array $defaultConfig = array(), GatewayFactoryInterface $coreGatewayFactory = null)
     {
-        $this->corePaymentFactory = $corePaymentFactory ?: new CorePaymentFactory();
+        $this->coreGatewayFactory = $coreGatewayFactory ?: new CoreGatewayFactory();
         $this->defaultConfig = $defaultConfig;
     }
 
@@ -34,7 +34,7 @@ class PaymentFactory implements PaymentFactoryInterface
      */
     public function create(array $config = array())
     {
-        return $this->corePaymentFactory->create($this->createConfig($config));
+        return $this->coreGatewayFactory->create($this->createConfig($config));
     }
 
     /**
@@ -44,7 +44,7 @@ class PaymentFactory implements PaymentFactoryInterface
     {
         $config = ArrayObject::ensureArrayObject($config);
         $config->defaults($this->defaultConfig);
-        $config->defaults($this->corePaymentFactory->createConfig((array) $config));
+        $config->defaults($this->coreGatewayFactory->createConfig((array) $config));
 
         $config->defaults(array(
             'payum.factory_name' => 'jms_payment_plugin',
